requirements.txt
factory-boy==3.3.0
Faker==24.9.0
python-dateutil==2.9.0.post0
six==1.16.0
